% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bedrock}
\alias{bedrock}
\title{Emission factors for bedrock types}
\format{
A data frame with 86 rows and 15 variables:
}
\usage{
bedrock
}
\description{
Based on SDCA research
}
\keyword{datasets}
