% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_calcualtion.R
\name{cap_gradient}
\alias{cap_gradient}
\title{Allocate cut and fill with maximum gradient}
\usage{
cap_gradient(raster_data, max_gradient = 1.5)
}
\arguments{
\item{raster_data}{a sf dataframe of points and data from the raster}

\item{max_gradient}{maximum gradient as %}
}
\value{
a dataframe of heights
}
\description{
extract elevations of a line and work our road gradient
}
