% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{climate_zones}
\alias{climate_zones}
\title{Warm Temperate Wet Climate Zone boundary}
\format{
A sf data.frame with one muliploygon
}
\usage{
climate_zones
}
\description{
Based on SDCA research
Approximate Warm Temperate Wet Climate Zone boundary in the UK
}
\keyword{datasets}
