% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_calcualtion.R
\name{cut_fill}
\alias{cut_fill}
\title{Infer cut and fill}
\usage{
cut_fill(elevations)
}
\arguments{
\item{elevations}{a dataframe of elevation data}
}
\value{
a dataframe of cut and fill.
}
\description{
extract elevations of a line
}
\examples{
\dontrun{

}
}
