% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_rasters.R
\name{download_rasters}
\alias{download_rasters}
\title{Download Rasters for local use}
\usage{
download_rasters()
}
\description{
When running the package locally you need the raster data. This function
download the data and stores it in the package folder.
}
\examples{
\dontrun{
download_rasters()
}
}
