% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{emissions_factors}
\alias{emissions_factors}
\title{Assumed emission factors of modes of transport}
\format{
A data frame with 1 rows and 8 variables:
}
\source{
\url{https://www.gov.uk/government/publications/greenhouse-gas-reporting-conversion-factors-2021}
}
\usage{
emissions_factors
}
\description{
Taken from the 2021 DEFRA emission factors, emissions in kgco2e/km
}
\keyword{datasets}
