% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_travel_demand.R
\name{estimate_travel_demand}
\alias{estimate_travel_demand}
\title{Estimate Travel Demand}
\usage{
estimate_travel_demand(
  infra,
  desire,
  stop_buff_dist = 3000,
  infra_buff_dist_ratio = 7
)
}
\arguments{
\item{infra}{a sf dataframe of infrastucutre}

\item{desire}{as sf dataframe of desire lines}

\item{stop_buff_dist}{distance in metres to buffer stops}

\item{infra_buff_dist_ratio}{distance in metres to buffer infrastructure}
}
\value{
a dataframe of
}
\description{
Estimate the travel deamnd from desire lines
}
