% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_calcualtion.R
\name{extract_elevations}
\alias{extract_elevations}
\title{Extract elevations}
\usage{
extract_elevations(line, path_dem)
}
\arguments{
\item{line}{a sf dataframe of linestrings}

\item{path_dem}{path to dem}
}
\value{
a dataframe of heights
}
\description{
extract elevations of a line
}
\examples{
\dontrun{

}
}
