% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_calcualtion.R
\name{extract_rasters}
\alias{extract_rasters}
\title{Extract data from rasters}
\usage{
extract_rasters(
  infra,
  rast_dem,
  rast_bedrock,
  rast_superficial,
  rast_landcover
)
}
\arguments{
\item{infra}{a sf dataframe}

\item{rast_dem}{stars raster of  dem}

\item{rast_bedrock}{stars raster of  bedrock}

\item{rast_superficial}{stars raster of  superficial}

\item{rast_landcover}{stars raster of  landcover}
}
\value{
a dataframe of heights
}
\description{
extract elevations of a line
}
