% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_geojson_to_api.R
\name{geojson_api}
\alias{geojson_api}
\title{Send to API}
\usage{
geojson_api(
  path,
  url = "https://dev.carbon.place/api/v1/locations.json?geojson=",
  showinput = TRUE
)
}
\arguments{
\item{path}{path to geojson}

\item{url}{API url}

\item{showinput}{logical, if TRUE input to R package is returned else resutls
are returned}
}
\value{
a named list of data frames and other objects
}
\description{
Function that take JSON from the API and convert into R format
}
\examples{
\dontrun{
parse_json("jsonhere")
}
}
