% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{landcover_factors}
\alias{landcover_factors}
\title{Emission factors for change in land use which match the CORINE land cover map}
\format{
A data frame with 37 rows and 4 variables:
}
\usage{
landcover_factors
}
\description{
Based on SDCA research
}
\details{
Factors in kgCo2 per m2 of land changed to settlement (i.e. built on)
Two environmental zones included cold_temp_wet  and warm_temp_wet
UK is mostly cold_temp_wet except for south west and London.
}
\keyword{datasets}
