% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_travel_demand.R
\name{line_to_bearing}
\alias{line_to_bearing}
\title{Convert lines to bearings}
\usage{
line_to_bearing(l, simplify = TRUE)
}
\arguments{
\item{l}{linestring}

\item{simplify}{logical, make all positive and rotate 90 degrees}
}
\value{
numeric between -pi and +pi
}
\description{
Only works for straight lines
}
