% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_travel_demand.R
\name{linestring_to_line}
\alias{linestring_to_line}
\title{Convert linestring into straight line}
\usage{
linestring_to_line(l)
}
\arguments{
\item{l}{linestring}
}
\value{
a sf dataframe
}
\description{
Convert linestring into straight line
}
