% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measure_infrastucture.R
\name{measure_infrastucture}
\alias{measure_infrastucture}
\title{Measure Infrastructure characteristics}
\usage{
measure_infrastucture(
  infra,
  path_dem = "D:/GitHub/SDCA-tool/sdca-data-prep/data/UK-dem-50m-4326-Int16.tif"
)
}
\arguments{
\item{infra}{a sf dataframe of infrastructure}

\item{path_dem}{path to the DEM raster}
}
\value{
a dataframe of
}
\description{
Measre infrastrcutre dimensions
}
\examples{
\dontrun{

}
}
