% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mode_shifts}
\alias{mode_shifts}
\title{Assumed mode shift for types of infrastructure}
\format{
A data frame with 6 rows and 9 variables:
}
\source{
\url{}
}
\usage{
mode_shifts
}
\description{
Based on SDCA research
}
\keyword{datasets}
