% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_json.R
\name{parse_json}
\alias{parse_json}
\title{Parse JSON}
\usage{
parse_json(json)
}
\arguments{
\item{json}{a json string as a character}
}
\value{
a named list of data frames and other objects
}
\description{
Function that take JSON from the API and convert into R format
}
\examples{
\dontrun{
parse_json("jsonhere")
}
}
