% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_results.R
\name{process_results}
\alias{process_results}
\title{Process results}
\usage{
process_results(args)
}
\arguments{
\item{args}{a json string as a character}
}
\value{
a json string as a character
}
\description{
Main function that the webtool uses to process resutls from the API
}
\examples{
\dontrun{
process_results("jsonhere")
}
}
