% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_VISUM.R
\name{read_visum}
\alias{read_visum}
\title{Read VISUM matrix files}
\usage{
read_visum(x)
}
\arguments{
\item{x}{a file path}
}
\description{
Import the v type VISUM matrix
}
\examples{
\dontrun{
test_function(2)
}
}
