% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{superficial}
\alias{superficial}
\title{Emission factors for superficial geology}
\format{
A data frame with 8 rows and 16 variables:
}
\usage{
superficial
}
\description{
Based on SDCA research
}
\keyword{datasets}
