% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function}
\alias{test_function}
\title{Test Function}
\usage{
test_function(x)
}
\arguments{
\item{x}{a number}
}
\description{
will square any number
}
\examples{
\dontrun{
test_function(2)
}
}
