% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_function.R
\name{test_function_fail}
\alias{test_function_fail}
\title{Test Function that fails}
\usage{
test_function_fail(x)
}
\arguments{
\item{x}{a number}
}
\description{
this function always fails
}
\examples{
\dontrun{
test_function_fail(2)
}
}
