% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_calcualtion.R
\name{cap_gradient}
\alias{cap_gradient}
\title{Allocate cut and fill with maximum gradient}
\usage{
cap_gradient(line, path_dem, max_gradient = 1.5)
}
\arguments{
\item{line}{a sf dataframe of linestrings}

\item{path_dem}{path to dem}

\item{max_gradient}{maximum gradient as %}
}
\value{
a dataframe of heights
}
\description{
Allocate cut and fill with maximum gradient
}
