% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_calcualtion.R
\name{cap_gradient}
\alias{cap_gradient}
\title{Infer profile of new road}
\usage{
cap_gradient(raster_data, max_gradient = 1.5)
}
\arguments{
\item{raster_data}{a dataframe of from `extract_rasters()`}

\item{max_gradient}{max gradient as percentage}
}
\value{
a dataframe with extra columns
}
\description{
extract elevations of a line with max gradient
}
