% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_calcualtion.R
\name{cut_fill}
\alias{cut_fill}
\title{Infer cut and fill}
\usage{
cut_fill(raster_data, width = 19)
}
\arguments{
\item{raster_data}{a dataframe of from `extract_rasters()`}

\item{width}{width of infrastrucutre}
}
\value{
a dataframe of cut and fill.
}
\description{
extract elevations of a line
}
