% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_travel_demand.R
\name{estimate_travel_demand}
\alias{estimate_travel_demand}
\title{Estimate Travel Demand}
\usage{
estimate_travel_demand(infra, desire)
}
\arguments{
\item{infra}{a sf dataframe of infrastucutre}

\item{desire}{as sf dataframe of desire lines}
}
\value{
a dataframe of
}
\description{
Estimate the travel deamnd from desire lines
}
\examples{
\dontrun{

}
}
