% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_landuse.R
\name{evaluate_landcover}
\alias{evaluate_landcover}
\title{Evaluate emissions from change in land cover}
\usage{
evaluate_landcover(infra, path_landcover, width = 19)
}
\arguments{
\item{infra}{a sf dataframe of linestrings}

\item{width}{width of infrastructure assumed for lines}

\item{path_dem}{path to landcover map}
}
\value{
a total emissions estimate in kgCO2
}
\description{
Emissions produced by building over natural land
}
\examples{
\dontrun{

}
}
