% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_materials.R
\name{evaluate_materials}
\alias{evaluate_materials}
\title{Evaluate Materials}
\usage{
evaluate_materials(
  infra,
  intervention_assets,
  intervention_assets_parameters,
  asset_components,
  carbon_factors,
  material_sites
)
}
\arguments{
\item{infra}{data frame with infrastructure properties}

\item{intervention_assets}{a pre-subset data frame}

\item{intervention_assets_parameters}{a pre-subset data frame}

\item{asset_components}{a pre-subset data frame}

\item{carbon_factors}{a pre-subset data frame}

\item{material_sites}{a data frame of the nearest material sites}
}
\value{
a json string as a character
}
\description{
Process the four key tables and get carbon factors for materials
}
\examples{
\dontrun{
evaluate_materials()
}
}
