% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/send_geojson_to_api.R
\name{geojson_api}
\alias{geojson_api}
\title{Send to API}
\usage{
geojson_api(
  path,
  url = "https://dev.carbon.place/api/v1/locations.json?geojson=",
  showinput = TRUE,
  path_dem = paste0(.libPaths()[1], "/sdca/exdata/UKdem.tif"),
  path_landcover = paste0(.libPaths()[1], "/sdca/exdata/landcover.tif"),
  path_bedrock = paste0(.libPaths()[1], "/sdca/exdata/bedrock.tif"),
  path_superficial = paste0(.libPaths()[1], "/sdca/exdata/superficial.tif")
)
}
\arguments{
\item{path}{path to geojson downloaded from website}

\item{url}{API url}

\item{showinput}{logical, if TRUE input to R package is returned else resutls
are returned}

\item{path_dem}{path to the terrain raster, defaults to location set by `download_rasters()`}

\item{path_landcover}{path to the land cover raster, defaults to location set by `download_rasters()`}

\item{path_bedrock}{path to the bedrock raster, defaults to location set by `download_rasters()`}

\item{path_superficial}{path to the superficial raster, defaults to location set by `download_rasters()`}
}
\value{
a named list of data frames and other objects
}
\description{
Function that take JSON from the API and convert into R format
}
\examples{
\dontrun{
parse_json("jsonhere")
}
}
